Rapidity Calculator
This sketch calculates the player's button pressing rapidity. It uses a digital pin to read
analog voltage: the simple method to detect if the button has been pressed. It counts the
number of times the player has pressed the button for 5 seconds. Then it calculates the result
and sends the data via serial monitor. Finally it terminates itself until the arduino is reset.