Improved Simple Light Sensor
This sketch creates a simple light sensor with some enhanced features.
It mplements the basic mechanism to detect ambient light with a photoresistor.
Creating a voltage divider with photoresistor and a 1k resistor, it measures the
voltage across the 1k resistor. Then it calculates the resistance of the photoresistor and
total power consumption of the sensor with mathematical equations and sends
the data via serial monitor.

Features:
1.
The resistance of the photoresistor is sent via serial monitor in real-time.
This visualises how the resistance increases in the dark, how power consumption
decreases due to increased equivalent resistance and vice versa.

2.
Additionally, the sketch lights up an LED in the dark and turns that off automatically
when there's adequate ambient light. The same implementation can be used for
home automation. For instance, it can turn ON the backup light during power outage.